feat: refactor Input component to include masking functionality and remove InputMask component
