--- src/components/ui/input-mask.tsx
@@ -1,41 +1,0 @@
-import * as React from 'react'
-import { cn } from '@/lib/utils';
-
-type InputProps = React.ComponentProps<'input'> & {
-    mask?: (value: string) => string;
-};
-
-function InputMask({ className, type, mask, onChange, ...props }: InputProps) {
-    function handleChange(e: React.ChangeEvent<HTMLInputElement>) {
-        if (mask) {
-            const maskedValue = mask(e.target.value);
-            const newEvent = {
-                ...e,
-                target: {
-                    ...e.target,
-                    value: maskedValue,
-                },
-            };
-            onChange?.(newEvent as React.ChangeEvent<HTMLInputElement>);
-        } else {
-            onChange?.(e);
-        }
-    }
-
-    return (
-        <input
-            type={type}
-            data-slot="input"
-            className={cn(
-                'file:text-foreground placeholder:text-muted-foreground selection:bg-primary selection:text-primary-foreground dark:bg-input/30 flex h-12 w-full min-w-0 rounded-md border-border border bg-transparent px-3 py-1 text-base shadow-xs transition-[color,box-shadow] outline-none file:inline-flex file:h-7 file:border-0 file:bg-transparent file:text-sm file:font-medium disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50 md:text-sm',
-                'focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px]',
-                'aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive',
-                className,
-            )}
-            onChange={handleChange}
-            {...props}
-        />
-    );
-}
-
-export { InputMask };
