+++ src/stories/Input.stories.tsx
@@ -1,6 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 import { Input } from '@/components/ui/input'
 import '@/index.css'
+import { useState } from 'react'
 
 const meta = {
     title: 'UI/Input',
@@ -50,5 +51,30 @@ export const WithError: Story = {
     args: {
         placeholder: 'Invalid input',
         'aria-invalid': true,
+    },
+}
+
+export const PhoneMasked: Story = {
+    render: () => {
+        const [value, setValue] = useState('')
+
+        const maskPhone = (input: string) => {
+            const digits = input.replace(/\D/g, '').slice(0, 9)
+            const parts = digits.match(/^(\d{0,3})(\d{0,3})(\d{0,3})$/)
+            if (!parts) return digits
+            return [parts[1], parts[2], parts[3]].filter(Boolean).join(' ')
+        }
+
+        return (
+            <div style={{ width: 300 }}>
+                <Input
+                    type="tel"
+                    placeholder="623 623 623"
+                    value={value}
+                    onChange={(e) => setValue(e.target.value)}
+                    mask={maskPhone}
+                />
+            </div>
+        )
     },
 }
